fix with float 64

+++ sensors/include/CompassSensor.h
@@ -6,6 +6,7 @@
 #include <string>
 #include <sensor_msgs/Imu.h>
 #include <std_msgs/Float32.h>
+#include <std_msgs/Float64.h>
 
 class CompassSensor
 {
@@ -19,7 +20,7 @@ public:
 	/// Post: 
 	CompassSensor(){
 		ros::Subscriber compassSub = n.subscribe("/mavros/global_position/compass_hdg", 1, &CompassSensor::CompassSensorCB, this);
-		_anglePub = n.advertise<std_msgs::Float32>("/IARC/currentAngle", 1);
+		_anglePub = n.advertise<std_msgs::Float64>("/IARC/currentAngle", 1);
 	}
 	
 	void CompassSensorCB(const std_msgs::Float32::ConstPtr& msg){
